reorganize tests to consolidate them
